Update README and fix test
