+++ next.config.ts
@@ -2,7 +2,6 @@ import createNextIntlPlugin from 'next-intl/plugin';
  
 const withNextIntl = createNextIntlPlugin();
 
-/** @type {import('next').NextConfig} */
 const nextConfig = {
   i18n: {
     locales: ['en'],
