+++ src/components/Header/Header.tsx
@@ -1,0 +1,7 @@
+import { useTranslations } from "next-intl";
+
+export const Header = () => {
+  const t = useTranslations('HomePage');
+  const title = t('title');
+  return <div>Header title={title}</div>
+}
