+++ src/pages/Some/index.tsx
@@ -1,12 +1,10 @@
 import Head from "next/head";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
 import {useTranslations} from 'next-intl';
-import styles from './Register.module.css';
-import { FormRegister } from "@/components/Form/Implementation/FormRegister/FormRegister";
-const queryClient = new QueryClient();
 
 
-export default function Register() {
+
+export default function Some() {
   const t = useTranslations('Register');
   const T_title = t('title');
 
@@ -15,10 +13,7 @@ export default function Register() {
       <Head>
         <title>{T_title}</title>
       </Head>
-      <QueryClientProvider client={queryClient}>
-        <h1 className={styles.h1}>Register</h1>
-        <FormRegister />
-      </QueryClientProvider>
+     
     </main>
   );
 }
